requirements.txt
pyspotlight==0.7.1
pandas==0.20.3
gensim==2.2.0
numpy==1.13.1